feat(friend-request): implement friend request creation logic
- Add FriendRequestRepository with Exists and Create methods for checking duplicates and persisting requests.
- Introduce FriendRequestService to encapsulate the friend request creation logic.
- Update variable naming: use 'receiverID' instead of 'requestedID' for clarity.
- Integrate service logic with handler to set the requester and validate duplicate requests.

+++ internal/models/friend.go
@@ -5,8 +5,8 @@ import "github.com/jinzhu/gorm"
 // FriendRequest represents a pending (or resolved) friend request.
 type FriendRequest struct {
 	gorm.Model
-	RequesterID int    `json:"requester_id"`
-	ReceiverID  int    `json:"receiver_id"`
+	RequesterID int64  `json:"requester_id"`
+	ReceiverID  int64  `json:"receiver_id"`
 	Status      string `json:"status"`
 	Message     string `json:"message"`
 }
